+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - Dataset generation: CVSS are now extracted from GitHub and PySec security advisories.
+- Dataset generation: CVSS, CPE, title and description (summary) are now extracted from CSAF document.
 
 
 ## Release 1.1.0 (2025-02-27)